+++ services/GateWay/include/global/Global.hpp
@@ -60,6 +60,15 @@ constexpr const char* DB_USER = "root";       // 数据库用户名
 constexpr const char* DB_PASSWORD = "whx";    // 数据库密码
 constexpr const char* DB_NAME = "chatroom";   // 数据库名称
 constexpr std::size_t DB_MAX_POOL_SIZE = 16;  // 数据库最大连接池大小
+
+constexpr const char* REDIS_HOST = "127.0.0.1";  // Redis 主机地址
+constexpr std::uint16_t REDIS_PORT = 6379;       // Redis 端口
+constexpr const char* REDIS_PASSWORD = "whx";    // Redis 密码
+constexpr std::size_t REDIS_DB_INDEX = 0;        // Redis 数据库索引
+constexpr std::size_t REDIS_MAX_POOL_SIZE = 16;  // Redis 最大连接池大小
+constexpr std::size_t REDIS_TIMEOUT = 3;         // Redis 连接超时时间
+
+constexpr const char* VERIFY_CODE_PREFIX = "verify_code_";  // 验证码在 Redis 中的键前缀
 }  // namespace server
 
 }  // namespace global