+++ services/GateWay/src/utils/common/code.hpp
@@ -24,6 +24,8 @@ constexpr std::int16_t JSON_PARSE_ERROR = 4;        // JSON解析错误
 constexpr std::int16_t INVALID_EMAIL_FORMAT = 5;    // 邮箱格式无效
 constexpr std::int16_t SEND_EMAIL_CODE_FAILED = 6;  // 发送验证码失败
 constexpr std::int16_t DATABASE_ERROR = 7;          // 数据库错误
+constexpr std::int16_t REDIS_ERROR = 8;             // Redis错误
+constexpr std::int16_t CODE_ALREADY_SENT = 9;       // 验证码已发送
 
 }  // namespace utils